Rename a directory and delete an unused file
