--- client/utilities/general.js
@@ -1,4 +1,0 @@
-export const deCaseDeSpace = (string) => {
-  const spaces = /\s/g;
-  return string.toLowerCase().replace(spaces, '');
-};
